Fastapi-SQLite/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
click==8.1.8
dill==0.4.0
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
pip==25.0.1
platformdirs==4.3.7
pydantic==2.11.4
pydantic_core==2.33.2
pylint==3.3.6
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tomlkit==0.13.2
typing_extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2